+++ apps/library/tests/factories/__init__.py
@@ -1,0 +1,18 @@
+from .author import AuthorFactory
+from .masterclass import MasterClassFactory
+from .participationapplication import ParticipationApplicationFestivalFactory
+from .performance import PerformanceFactory
+from .play import PlayFactory, ProgramFactory
+from .reading import ReadingFactory
+from .team_member import TeamMemberFactory
+
+__all__ = (
+    AuthorFactory,
+    MasterClassFactory,
+    ParticipationApplicationFestivalFactory,
+    PerformanceFactory,
+    PlayFactory,
+    ProgramFactory,
+    ReadingFactory,
+    TeamMemberFactory,
+)
